perf: Import plugin-resolver PEX rules lazily
